Print out destructor
